requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
numpy==2.1.2
openai==1.52.2
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sniffio==1.3.1
tabulate==0.9.0
tk==0.1.0
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
